+++ js/data.js
@@ -1,0 +1,36 @@
+const DESCRIPTIONS = [
+  'Lorem ipsum',
+  'dolor sit',
+  'amet consectetur',
+  'adipisicing elit',
+  'Accusamus, esse',
+  'error ipsam',
+  'voluptas ab',
+  'eveniet unde',
+  'exercitationem fugiat',
+  'quis eum',
+  'iste perferendis',
+  'ipsum earum iure!',
+];
+
+const COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const NAMES = [
+  'Джереми',
+  'Жозефина',
+  'Антуанетта',
+  'Крэг Хек',
+  'Гервульф',
+  'Лабета',
+  'Луна',
+  'Джейсон Стэтхэм',
+];
+
+export { DESCRIPTIONS, COMMENTS, NAMES };
